reorganise to make networkx standard
